* Restored spell casting after the first one * The whole update system now runs on microseconds instead of milliseconds

+++ gamed/include/Champion.h
@@ -21,7 +21,7 @@ public:
    Spell* castSpell(uint8 slot, float x, float y, Unit* target);
    Spell* levelUpSpell(uint8 slot);
    
-   virtual void update(unsigned int diff);
+   virtual void update(int64 diff);
    
    uint8 getSkillPoints() const { return skillPoints; }
